move Connection/EventSubClient to its own file
