+++ src/providers/twitch/eventsub/Controller.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "providers/twitch/eventsub/SubscriptionRequest.hpp"
-#include "twitch-eventsub-ws/listener.hpp"
 #include "twitch-eventsub-ws/session.hpp"
 
 #include <boost/asio/executor_work_guard.hpp>
@@ -15,48 +14,6 @@
 #include <thread>
 
 namespace chatterino::eventsub {
-
-class EventSubClient final : public lib::Listener
-{
-public:
-    void onSessionWelcome(
-        lib::messages::Metadata metadata,
-        lib::payload::session_welcome::Payload payload) override;
-
-    void onNotification(lib::messages::Metadata metadata,
-                        const boost::json::value &jv) override;
-
-    void onChannelBan(lib::messages::Metadata metadata,
-                      lib::payload::channel_ban::v1::Payload payload) override;
-
-    void onStreamOnline(
-        lib::messages::Metadata metadata,
-        lib::payload::stream_online::v1::Payload payload) override;
-
-    void onStreamOffline(
-        lib::messages::Metadata metadata,
-        lib::payload::stream_offline::v1::Payload payload) override;
-
-    void onChannelChatNotification(
-        lib::messages::Metadata metadata,
-        lib::payload::channel_chat_notification::v1::Payload payload) override;
-
-    void onChannelUpdate(
-        lib::messages::Metadata metadata,
-        lib::payload::channel_update::v1::Payload payload) override;
-
-    void onChannelChatMessage(
-        lib::messages::Metadata metadata,
-        lib::payload::channel_chat_message::v1::Payload payload) override;
-
-    QString getSessionID() const
-    {
-        return this->sessionID;
-    }
-
-private:
-    QString sessionID;
-};
 
 class Controller
 {
